compilador_simples/requirements.txt
antlr4-python3-runtime==4.13.1
llvmlite>=0.40.0